Fixed key required error
This wasn't working because I had a separate function that returned the hiccup component that the key associates with, when I moved the separate function logic directly into this function it works properly.

Also removed the wrapping ul element I had kept because of a css class I was using for dumb reasons that was inheriting css hover events. So you'd hover over the title and it would do the hover animation for list elements which was dumb.

Also updated the title to display the title of the episode. Should've done this in a separate branch but whatever. I wanted to do it quick and it's just me, lol. I can't step on my own toes.